Removing explicit casting from line 44

+++ ft_atoi.c
@@ -41,7 +41,7 @@ int	ft_atoi(const char *str)
 	}
 	while (ft_isdigit(str[i]))
 	{
-		result = result * 10 + (int) str[i] - 48;
+		result = result * 10 + str[i] - 48;
 		i++;
 	}
 	return (result * sign);
